Restructure the custom functions
Removed all functions from functions.php file and moved them to a new folder called includes.

All custom functions will be available at this folder, importing them inside functions.php through a require_once function.
