+++ includes/displays-url-params-on-frontend.php
@@ -1,0 +1,14 @@
+<?php
+
+/*
+ * Captures parameters from URL and shows it inside a shortcode on front-end
+ */
+function URLParam($atts)
+{
+    extract(shortcode_atts(array(
+        'param' => 'param',
+    ), $atts));
+    return $_GET[$param];
+}
+add_shortcode('URLParam', 'URLParam');
+
